🧩 Refactor(translation).
