--- kitx_mobile/lib/utils/translation.dart
@@ -1,140 +1,0 @@
-import 'package:get/get.dart';
-
-/// Translation Data
-class Translation extends Translations {
-  @override
-  Map<String, Map<String, String>> get keys => {
-        'zh_CN': {
-          'Public_Enable': '启用',
-          'Public_Disable': '禁用',
-          'Public_Animation': '动画',
-          'Public_Additional': '额外的',
-          'Public_Log': '日志',
-          'Public_Reload': '重载',
-          'Public_Restart': '重启',
-          'Public_Chat': '聊天',
-          'Public_Closed': "已关闭",
-          'Public_Error': "错误",
-          'Public_Launch': "启动",
-          'Public_Launching': "启动中",
-          'Public_Stop': "停止",
-          'Public_Stopping': "停止中",
-          'ServiceStatus_Error': '发生错误',
-          'ServiceStatus_Starting': '正在启动',
-          'ServiceStatus_Stopping': '正在停止',
-          'ServiceStatus_Running': '正在运行',
-          'ServiceStatus_Pending': '已停止',
-          'NotificationService_StatusNotificationTitle': 'KitX %status%',
-          'NotificationService_StatusNotificationBody': '当前设备数: %device_count%',
-          'Drawer_Title': 'KitX',
-          'Drawer_Devices': '设备管理',
-          'Drawer_Plugins': '插件列表',
-          'Drawer_Account': '账户',
-          'Drawer_Test': '测试',
-          'Drawer_Setting': '设置',
-          'Drawer_About': '关于',
-          'Option_RestartDevicesServer': '重启设备网络服务器',
-          'Option_ShutdownDevicesServer': '关闭设备网络服务器',
-          'IndexPage_Title': 'KitX',
-          'HomePage_DevicesCount': '设备在线',
-          'DevicesPage_Title': '设备管理',
-          'DevicesPage_PluginsCountText': '%count% 插件在线',
-          'DevicesPage_LocalDevice': '(本机)',
-          'DevicesPage_MainDevice': '(主控)',
-          'DeviceChatPage_Title': '聊天',
-          'PluginsPage_Title': '插件列表',
-          'PluginsPage_BuiltInPlugins': '内置插件',
-          'PluginsPage_ConnectedPlugins': '已连接的插件',
-          'AccountPage_Title': '账户',
-          'TestPage_Title': '测试',
-          'SettingsPage_Title': '设置',
-          'SettingsPage_Languages': '语言',
-          'SettingsPage_Theme': '主题',
-          'SettingsPage_Light': '浅',
-          'SettingsPage_Dark': '深',
-          'SettingsPage_FollowSystem': '跟随',
-          'SettingsPage_CleanLog': '清理日志',
-          'SettingsPage_Saved': '配置已保存!',
-          'AboutPage_Title': '关于',
-          'AboutPage_Version': '版本',
-          'AboutPage_VersionCode': '构建号',
-          'AboutPage_Contributors': '贡献者',
-          'AboutPage_Repos': '仓库',
-          'AboutPage_ThirdPartyLicenses': '第三方许可',
-          'AboutPage_ThirdPartyLicenses_DisplayAll': '显示全部',
-          'AboutPage_JoinUS': '加入我们',
-          'AboutPage_JoinUS_Text': ''
-              '我们的项目都托管在 https://github.com/Crequency/, 欢迎 Fork 并贡献您宝贵的意见.'
-              '\n\n'
-              'KitX Mobile 的源码全部位于 https://github.com/Crequency/KitX-Mobile/'
-              '\n\n'
-              'KitX Mobile 是 KitX Project 一部分, KitX Project 的主仓库是 https://github.com/Crequency/KitX',
-        },
-        'en_US': {
-          'Public_Enable': 'Enable',
-          'Public_Disable': 'Disable',
-          'Public_Animation': 'Animation',
-          'Public_Additional': 'Additional',
-          'Public_Log': 'Log',
-          'Public_Reload': 'Reload',
-          'Public_Restart': 'Restart',
-          'Public_Chat': 'Chat',
-          'Public_Closed': "Closed",
-          'Public_Error': "Error",
-          'Public_Launch': "Launch",
-          'Public_Launching': "Launching",
-          'Public_Stop': "Stop",
-          'Public_Stopping': "Stopping",
-          'ServiceStatus_Error': 'Error',
-          'ServiceStatus_Starting': 'Starting',
-          'ServiceStatus_Stopping': 'Stopping',
-          'ServiceStatus_Running': 'Running',
-          'ServiceStatus_Pending': 'Stopped',
-          'NotificationService_StatusNotificationTitle': 'KitX %status%',
-          'NotificationService_StatusNotificationBody': '%device_count% Devices Online',
-          'Drawer_Title': 'KitX',
-          'Drawer_Devices': 'Devices',
-          'Drawer_Plugins': 'Plugins',
-          'Drawer_Account': 'Account',
-          'Drawer_Test': 'Test',
-          'Drawer_Setting': 'Settings',
-          'Drawer_About': 'About',
-          'Option_RestartDevicesServer': 'Restart Devices Server',
-          'Option_ShutdownDevicesServer': 'Shutdown Devices Server',
-          'IndexPage_Title': 'KitX',
-          'HomePage_DevicesCount': 'Devices found.',
-          'DevicesPage_Title': 'Devices',
-          'DevicesPage_PluginsCountText': '%count% Plugins Online',
-          'DevicesPage_LocalDevice': '(Local)',
-          'DevicesPage_MainDevice': '(Master)',
-          'DeviceChatPage_Title': 'Chat',
-          'PluginsPage_Title': 'Plugins',
-          'PluginsPage_BuiltInPlugins': 'Built-in Plugins',
-          'PluginsPage_ConnectedPlugins': 'Connected Plugins',
-          'AccountPage_Title': 'Account',
-          'TestPage_Title': 'Test',
-          'SettingsPage_Title': 'Settings',
-          'SettingsPage_Languages': 'Languages',
-          'SettingsPage_Theme': 'Theme',
-          'SettingsPage_Light': 'Light',
-          'SettingsPage_Dark': 'Dark',
-          'SettingsPage_FollowSystem': 'Follow system',
-          'SettingsPage_CleanLog': 'Clean Log',
-          'SettingsPage_Saved': 'Changes saved!',
-          'AboutPage_Title': 'About',
-          'AboutPage_Version': 'Version',
-          'AboutPage_VersionCode': 'Build Number',
-          'AboutPage_Contributors': 'Contributors',
-          'AboutPage_Repos': 'Repos',
-          'AboutPage_ThirdPartyLicenses': 'Third-Party Licenses',
-          'AboutPage_ThirdPartyLicenses_DisplayAll': 'Display all',
-          'AboutPage_JoinUS': 'Join US',
-          'AboutPage_JoinUS_Text': ''
-              'All our projects storage on https://github.com/Crequency/, welcome to fork and contribute to them.'
-              '\n\n'
-              'Source code of KitX Mobile located at https://github.com/Crequency/KitX-Mobile/'
-              '\n\n'
-              'KitX Mobile belongs to KitX Project, Main repo of KitX Project located at https://github.com/Crequency/KitX',
-        }
-      };
-}
