+++ kitx_mobile/lib/utils/translation/en_us.dart
@@ -1,0 +1,66 @@
+// ignore_for_file: constant_identifier_names
+
+/// English - United States (en_US) translation
+const Map<String, String> en_us = {
+  'Public_Enable': 'Enable',
+  'Public_Disable': 'Disable',
+  'Public_Animation': 'Animation',
+  'Public_Additional': 'Additional',
+  'Public_Log': 'Log',
+  'Public_Reload': 'Reload',
+  'Public_Restart': 'Restart',
+  'Public_Chat': 'Chat',
+  'Public_Closed': 'Closed',
+  'Public_Error': 'Error',
+  'Public_Launch': 'Launch',
+  'Public_Launching': 'Launching',
+  'Public_Stop': 'Stop',
+  'Public_Stopping': 'Stopping',
+  'Public_Running': 'Running',
+  'Public_Pending': 'Pending',
+  'NotificationService_StatusNotificationTitle': 'KitX %status%',
+  'NotificationService_StatusNotificationBody': '%device_count% Devices Online',
+  'Drawer_Title': 'KitX',
+  'Drawer_Devices': 'Devices',
+  'Drawer_Plugins': 'Plugins',
+  'Drawer_Account': 'Account',
+  'Drawer_Test': 'Test',
+  'Drawer_Setting': 'Settings',
+  'Drawer_About': 'About',
+  'Option_RestartDevicesServer': 'Restart Devices Server',
+  'Option_ShutdownDevicesServer': 'Shutdown Devices Server',
+  'IndexPage_Title': 'KitX',
+  'HomePage_DevicesCount': 'Devices found.',
+  'DevicesPage_Title': 'Devices',
+  'DevicesPage_PluginsCountText': '%count% Plugins Online',
+  'DevicesPage_LocalDevice': '(Local)',
+  'DevicesPage_MainDevice': '(Master)',
+  'DeviceChatPage_Title': 'Chat',
+  'PluginsPage_Title': 'Plugins',
+  'PluginsPage_BuiltInPlugins': 'Built-in Plugins',
+  'PluginsPage_ConnectedPlugins': 'Connected Plugins',
+  'AccountPage_Title': 'Account',
+  'TestPage_Title': 'Test',
+  'SettingsPage_Title': 'Settings',
+  'SettingsPage_Languages': 'Languages',
+  'SettingsPage_Theme': 'Theme',
+  'SettingsPage_Light': 'Light',
+  'SettingsPage_Dark': 'Dark',
+  'SettingsPage_FollowSystem': 'Follow system',
+  'SettingsPage_CleanLog': 'Clean Log',
+  'SettingsPage_Saved': 'Changes saved!',
+  'AboutPage_Title': 'About',
+  'AboutPage_Version': 'Version',
+  'AboutPage_VersionCode': 'Build Number',
+  'AboutPage_Contributors': 'Contributors',
+  'AboutPage_Repos': 'Repos',
+  'AboutPage_ThirdPartyLicenses': 'Third-Party Licenses',
+  'AboutPage_ThirdPartyLicenses_DisplayAll': 'Display all',
+  'AboutPage_JoinUS': 'Join US',
+  'AboutPage_JoinUS_Text': ''
+      'All our projects storage on https://github.com/Crequency/, welcome to fork and contribute to them.'
+      '\n\n'
+      'Source code of KitX Mobile located at https://github.com/Crequency/KitX-Mobile/'
+      '\n\n'
+      'KitX Mobile belongs to KitX Project, Main repo of KitX Project located at https://github.com/Crequency/KitX',
+};
