+++ kitx_mobile/lib/utils/translation/zh_cn.dart
@@ -1,0 +1,66 @@
+// ignore_for_file: constant_identifier_names
+
+/// Simplified Chinese (zh_CN) translation
+const Map<String, String> zh_cn = {
+  'Public_Enable': '启用',
+  'Public_Disable': '禁用',
+  'Public_Animation': '动画',
+  'Public_Additional': '额外的',
+  'Public_Log': '日志',
+  'Public_Reload': '重载',
+  'Public_Restart': '重启',
+  'Public_Chat': '聊天',
+  'Public_Closed': '已关闭',
+  'Public_Error': '错误',
+  'Public_Launch': '启动',
+  'Public_Launching': '启动中',
+  'Public_Stop': '停止',
+  'Public_Stopping': '停止中',
+  'Public_Running': '运行中',
+  'Public_Pending': '等待中',
+  'NotificationService_StatusNotificationTitle': 'KitX %status%',
+  'NotificationService_StatusNotificationBody': '当前设备数: %device_count%',
+  'Drawer_Title': 'KitX',
+  'Drawer_Devices': '设备管理',
+  'Drawer_Plugins': '插件列表',
+  'Drawer_Account': '账户',
+  'Drawer_Test': '测试',
+  'Drawer_Setting': '设置',
+  'Drawer_About': '关于',
+  'Option_RestartDevicesServer': '重启设备网络服务器',
+  'Option_ShutdownDevicesServer': '关闭设备网络服务器',
+  'IndexPage_Title': 'KitX',
+  'HomePage_DevicesCount': '设备在线',
+  'DevicesPage_Title': '设备管理',
+  'DevicesPage_PluginsCountText': '%count% 插件在线',
+  'DevicesPage_LocalDevice': '(本机)',
+  'DevicesPage_MainDevice': '(主控)',
+  'DeviceChatPage_Title': '聊天',
+  'PluginsPage_Title': '插件列表',
+  'PluginsPage_BuiltInPlugins': '内置插件',
+  'PluginsPage_ConnectedPlugins': '已连接的插件',
+  'AccountPage_Title': '账户',
+  'TestPage_Title': '测试',
+  'SettingsPage_Title': '设置',
+  'SettingsPage_Languages': '语言',
+  'SettingsPage_Theme': '主题',
+  'SettingsPage_Light': '浅',
+  'SettingsPage_Dark': '深',
+  'SettingsPage_FollowSystem': '跟随',
+  'SettingsPage_CleanLog': '清理日志',
+  'SettingsPage_Saved': '配置已保存!',
+  'AboutPage_Title': '关于',
+  'AboutPage_Version': '版本',
+  'AboutPage_VersionCode': '构建号',
+  'AboutPage_Contributors': '贡献者',
+  'AboutPage_Repos': '仓库',
+  'AboutPage_ThirdPartyLicenses': '第三方许可',
+  'AboutPage_ThirdPartyLicenses_DisplayAll': '显示全部',
+  'AboutPage_JoinUS': '加入我们',
+  'AboutPage_JoinUS_Text': ''
+      '我们的项目都托管在 https://github.com/Crequency/, 欢迎 Fork 并贡献您宝贵的意见.'
+      '\n\n'
+      'KitX Mobile 的源码全部位于 https://github.com/Crequency/KitX-Mobile/'
+      '\n\n'
+      'KitX Mobile 是 KitX Project 一部分, KitX Project 的主仓库是 https://github.com/Crequency/KitX',
+};
